Add partial views

+++ MyLeasing.Web/Controllers/PropertyTypesController.cs
@@ -47,8 +47,6 @@ namespace MyLeasing.Web.Controllers
         }
 
         // POST: PropertyTypes/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name")] PropertyType propertyType)
@@ -79,8 +77,6 @@ namespace MyLeasing.Web.Controllers
         }
 
         // POST: PropertyTypes/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] PropertyType propertyType)
@@ -122,21 +118,18 @@ namespace MyLeasing.Web.Controllers
             }
 
             PropertyType propertyType = await _context.PropertyTypes
+                .Include(pt => pt.Properties)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (propertyType == null)
             {
                 return NotFound();
             }
 
-            return View(propertyType);
-        }
-
-        // POST: PropertyTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            PropertyType propertyType = await _context.PropertyTypes.FindAsync(id);
+            if(propertyType.Properties.Count > 0)
+            {
+                //TODO: message
+                return RedirectToAction(nameof(Index));
+            }
             _context.PropertyTypes.Remove(propertyType);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
